selenium integration with cucumber.
We have to implement these steps in Java in order to test the feature files.We Need to create a class which contains those given, when and then statements. Cucumber uses its annotations and all the steps are embedded in those annotations (given, when, then).Each phrase starts with “^” so that cucumber understands the start of the step. Similarly, each step ends with “$”. The user can use regular expressions to pass different test data.

illustrate how feature files can be implemented.

In this example, I have not used any selenium API. This is to just show how cucumber works as standalone framework